refactor:♻️ConvertAggregatedToClustersをh3utilに移動してクリーンアーキテクチャ違反を解消(#6)
